removing/disabling more references to JobStepModel
