+++ src/SFA.DAS.Payments.Monitoring.Jobs.Application/Handlers/StartedProcessingEarningsEventHandler.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using NServiceBus;
 using SFA.DAS.Payments.Application.Infrastructure.Logging;
+using SFA.DAS.Payments.Monitoring.Jobs.Application.Services.Earnings;
 using SFA.DAS.Payments.Monitoring.Jobs.Messages.Commands;
 
 namespace SFA.DAS.Payments.Monitoring.Jobs.Application.Handlers
