+++ src/SFA.DAS.Payments.Monitoring.Jobs.Application/SFA.DAS.Payments.Monitoring.Jobs.Application.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="Autofac" Version="4.8.1" />
+    <PackageReference Include="AutoMapper" Version="7.0.1" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.1" />
     <PackageReference Include="NServiceBus" Version="7.1.5" />
   </ItemGroup>
