+++ src/SFA.DAS.Payments.Monitoring.Jobs.Application/Services/Batch/BatchQueue.cs
@@ -1,0 +1,7 @@
+﻿namespace SFA.DAS.Payments.Monitoring.Jobs.Application.Services.Batch
+{
+    public class BatchQueue
+    {
+        
+    }
+}
