+++ src/SFA.DAS.Payments.Monitoring.Jobs.Data/Model/JobStepModel.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace SFA.DAS.Payments.Monitoring.Jobs.Data.Model
+{
+    public class JobStepModel
+    {
+        public long Id { get; set; }
+        public long JobId { get; set; }
+        public Guid MessageId { get; set; }
+        public Guid? ParentMessageId { get; set; }
+        public JobMessageStatus Status { get; set; }
+        public DateTimeOffset? StartTime { get; set; }
+        public DateTimeOffset? EndTime { get; set; }
+        public string MessageName { get; set; }
+        public virtual JobModel Job { get; set; }
+    }
+}
